% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a state map for the state for a single data file}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state number, an integer value, this is the state for which data
will be mapped.}

\item{year}{an integer value for the year for which we want to plot
the location of the fatalities on the map.}
}
\value{
If the state number as mentioned in the parameter \code{state.num} is
        invalid, this function returns an error. But, if the \code(state.num)
        is valid, this function plots  the locations (using
        latitude and longitude variables) of the fatalities.
}
\description{
Map the fatalities (where the accident occured) for the 'state' given 
the dataset corresponding to the \code{year}.
}
\note{
This function creates a state map for a single state in the United states.
}
\examples{
fars_map_state(state.num = 12, year = 2012)

}
