% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{read data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(or filepath) name of the file to be read in.}
}
\value{
a tibble which contains the read in data.
}
\description{
Reads in data file (if the file exists) and then coerces
the data object into a tibble.
}
\examples{
fars_read("farsDataFile")
fars_read("path/to/file")

}
