% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Add a year column to data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list/vector containing the \code{years}, for which we need to
append the data file with a year column.}
}
\value{
Returns a list (if the year is valid, if invalid, returns NULL)
        whose each element is a tibble (with a year column added to it)
        corresponding to each year specified in the input \code{years}
        list/vector to the function.
}
\description{
There exists a single data file for each year. Not all years have 
data files. For those years that do have a data file, this function
adds a year column to that data file. For those years that don't have
a data file this function returns NULL.
}
\examples{
fars_read_years(years = list(2012, 1988, 2013, 2015))
fars_read_years(years = c(2012,  2013, 1964, 2000))
        
}
