% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{combine all data files and summarize the result}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{years a list/vector containing the \code{years}, for which we need to
append the data file with a year column.}
}
\value{
a summary tibble where each row represents a month and each column 
        represents a year. Each cell, represents the number of values 
        (the count) under each \code{month}, \code{year} combination.
}
\description{
Combine all data files (note that there is one file for each valid
year in the \code{years} vector/list) and then get a count for how many
values (the count) exists for each \code{month}, \code{year} combination.
The values represent the number of fatal injuries suffered in motor vehicle
traffic crashes.
}
