% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a file name string}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer value for the year in question}
}
\value{
a string for the file name that contains the year
        inputted by the user in the function.
}
\description{
Create a file name by specifying a specific year.
}
\examples{
make_filename(2012)
make_filename(1988)

}
