% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display a map of a state with the accidents plotted on the map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the number of a state as defined in the FARS data.}

\item{year}{An integer corresponding to the year of the FARS accident data.}
}
\value{
A figure of a state map with accidents plotted by geographic location
   as specified by latitude and longitude is printed.
}
\description{
This function creates a figure of a state map.  The points on the
map correspond to accidents in the FARS data.  Each accident is plotted
according to the latitude and longitude of the accident site.
}
\note{
If an invalid state number is used as input, the function stops with an
   error message.

If there are zero accidents for a particular state and year, a message
   is displayed and the function ends.
}
\examples{
\dontrun{
fars_map_state(1,2013)
}

}

