% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Number of a state included within the FARS dataset.}

\item{year}{A year (as YYYY) or vector of years.}
}
\value{
A map with accident locations for the given state and year.
}
\description{
Plots a map of a given state including the locations of accidents in the
given year or returns an error if state number is invalid or message if
there are no accidents for the selected state and year.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014,2015))
fars_summarize_years(2013:2015)
}

}
