% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string which is the name of the csv file
with the FARS data}
}
\value{
This function returns the FARS data in a tbl_df object
}
\description{
This is a function which reads the FARS data from a csv file
into a dataframe and converts it into a tbl_df object
}
\note{
If the filename does not exist, the function stops with an
   error message.
}
\examples{
filename<-system.file("extdata","accident_2013.csv.bz2",package="FARSanalysis")
fars_read(filename)

}

