% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string defining the name of a csv file in the 
working directory which the function will read.}
}
\value{
This function returns a tibble dataframe or an error if no filename
with the given name exists in the working directory.
}
\description{
A function which reads a US National Highway Traffic Safety Administration 
Fatality Analysis Reporting System (FARS) dataset and coerces it to a tibble.
}
\examples{
\dontrun{
fars_read(accident_2013.csv)
fars_read(accident_2014.csv.bz2)
}

}
