% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Select month and year data from files corresponding to a list of years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A integer for a single year or a vector of integers for a list of years.}
}
\value{
This function returns the MONTH and year data from a file for each year
   in the input list.  The data is returned as a list where each list item
   is a tbl_df object containing MONTH and year data for one file.
}
\description{
This function returns the MONTH and year data from each file from a list of
input files corresponding to a list of years.
}
\note{
This function returns an error if an invalid year is present
   in the input list.  An invalid year is one where the filename for that
   year does not exist.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(c(2013,2014))
}

}

