% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years (as YYYY).}
}
\value{
A list of tibble data frames for the input years with length equal
 to the input vector and containing only MONTH and year columns.
}
\description{
Reads MONTH and year variables from the input dataset and stores them as data
frames in a list.
}
\note{
An error will occur if year number is invalid.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(c(2013,2014,2015))
}

}
