% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize number of accidents per month of a year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A integer for a single year or a vector of integers for a list of years.}
}
\value{
This function returns the number of FARS accidents grouped by MONTH and year.
The data is returned as a tbl_df with a MONTH column followed by a
column corresponding to each year in the input list.
}
\description{
This function returns the number of accidents grouped by MONTH and year for
a given number of years.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014))
}

}

