% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years (as YYYY).}
}
\value{
A tibble data frame containing the number of accidents with months
as rows and years as columns.
}
\description{
Summarises the accidents by month and year for the tibble data frames
imported by \code{fars_read_years}.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(2013:2015)
}

}
