% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Integer corresponding to the year of the FARS accident data.}
}
\value{
This function returns a character string containing the filename
   of the accident data for a particular year.
}
\description{
This function returns a filename based on the year of the accident data.
}
\examples{
make_filename(2013)

}

