% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make FARS file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numerical year in the format YYYY.}
}
\value{
A string with the data file name for the corresponding input year and
 the file path within the package.
}
\description{
Generates a file name dependent on the given FARS file year.
}
\note{
Function does not first check if corresponding data file exists.
}
\examples{
 \dontrun{
 make_filename(2014)
 }
 
}
